+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Change log for the codebase. Initialised from the developments following version
 ### Added
 
 - added pydantic env
+- created initial pydantic dataModel
 
 ### Changed